refactor(logging): setup and CLI logger initialization
