+++ src/mixseek/cli/__init__.py
@@ -1,15 +1,11 @@
 """CLI module for MixSeek."""
 
-from mixseek.cli.output_logger import (
+from mixseek.observability.logging_setup import (
     get_cli_logger,
-    get_data_logger,
     setup_cli_logger,
-    setup_data_logger,
 )
 
 __all__ = [
     "get_cli_logger",
-    "get_data_logger",
     "setup_cli_logger",
-    "setup_data_logger",
 ]
