+++ src/mixseek/observability/__init__.py
@@ -15,7 +15,23 @@
 """
 
 from mixseek.observability.logfire import setup_logfire
-from mixseek.observability.logging_setup import get_log_format, is_logger_initialized, setup_logging
+from mixseek.observability.logging_setup import (
+    early_setup_cli_logger_from_env,
+    get_cli_logger,
+    get_log_format,
+    is_logger_initialized,
+    setup_cli_logger,
+    setup_logging,
+)
 from mixseek.observability.tee_writer import TeeWriter
 
-__all__ = ["get_log_format", "is_logger_initialized", "setup_logfire", "setup_logging", "TeeWriter"]
+__all__ = [
+    "TeeWriter",
+    "early_setup_cli_logger_from_env",
+    "get_cli_logger",
+    "get_log_format",
+    "is_logger_initialized",
+    "setup_cli_logger",
+    "setup_logfire",
+    "setup_logging",
+]
